+++ client/utility.hpp
@@ -19,7 +19,7 @@ namespace snek {
 
         [[nodiscard]] sf::Vector2f decode_vector(const std::string& vector_data);
 
-        [[nodiscard]] std::string encode_vector(const sf::Vector2f& vector);
+        [[nodiscard]] std::string encode_vector(const sf::Vector2i& vector);
 
         [[nodiscard]] std::vector<sf::Vector2f> decode_vector_array(const std::string& array_data);
     }
